Start on integration testing
