+++ rollup.config.js
@@ -1,7 +1,13 @@
 import commonjs from 'rollup-plugin-commonjs';
 import nodeResolve from 'rollup-plugin-node-resolve';
 
-const external = ['@codemirror/state', 'assert', 'ot-json0', 'ot-json1', ''];
+const external = [
+  '@codemirror/state',
+  'assert',
+  'ot-json0',
+  'ot-json1',
+  'jsdom',
+];
 const output = {
   format: 'cjs',
   sourcemap: true,
