--- src/codemirror.js
@@ -1,9 +1,0 @@
-// This file exists so that it's easy to change the package
-// from which CodeMirror imports originate.
-export {
-  EditorState,
-  StateField,
-  Transaction,
-  ChangeDesc,
-  ChangeSet,
-} from '@codemirror/state';
